Live-kit/pyproject.toml
[project]
name = "live-kit"
version = "0.1.0"
description = "Real-time voice agent with LiveKit"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "livekit-agents[deepgram,openai,silero,turn-detector,elevenlabs]~=1.2",
    "livekit-plugins-openai",
    "livekit-plugins-deepgram", 
    "livekit-plugins-elevenlabs",
    "livekit-plugins-silero",
    "livekit-plugins-turn-detector",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv",
    "redis>=5.0.0",
    "uuid",
    "google-generativeai>=0.3.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["agent"]
